feat: use smaller data types

+++ mandelbrot.h
@@ -1,4 +1,5 @@
 #include "mandelbrot_types.h"
+#include "stdlib.h"
 
 #ifndef MANDELBROT_H
 #define MANDELBROT_H
@@ -9,6 +10,6 @@ enum MandlebrotStrategy
     FOUR_SPLIT, // iteratively split into four planes
 };
 
-void calculate_mandelbrot(Zoom z, int max_iterations, int *result, enum MandlebrotStrategy strategy, int num_threads);
+void calculate_mandelbrot(Zoom z, uint16_t max_iterations, uint16_t *result, enum MandlebrotStrategy strategy, int num_threads);
 
 #endif
